Set current value and Lock/enable wallbox in home assistant
A slider to set the current limit will now show in home assistant.
There is also a switch to turn on/off the wallbox. This will set the current to 0 and will restore the previous current limit when turned on again. 
While the switch is turned off, the current limit can be changed, but the change will apply only after the wallbox is turned on.

Usefull for wallboxes that are mounted outside, so they could be used by other peoply while you are away