Enable/Disable Standby of HeidelbergWallbox (#49)
* Enable/Disable Standby of HeidelbergWallbox

Created a new MQTT topic /standby_enabled to control the Standby register of the Heidelberg Wallbox.
The topic is automatically discovered by Home Assistant.

Tested successfully on the LILYGO build:
- MQTT topic HeidelBridge/standby_enabled is published correctly.
- Home Assistant controls added.
- Standby mode can be set via Home Assistant → Wallbox enters deep standby after ~10-15 minutes.
- Wakeup via Heidelberg interface is not possible while in deep standby.
- Wallbox wakes up automatically when a car is connected, and control via Home Assistant/MQTT is restored.
- Disabling Standby also works.

EVCC will not work if Wallbox is in deep standby.

* Change IsStandbyEnabled() to GetStandbyEnabled()

Tried to implement suggentions of Giermann: https://github.com/BorisBrock/HeidelBridge/pull/49#issuecomment-3601021195

Renamed the method from IsStandbyEnabled() to GetStandbyEnabled() and using ReadRegisters to get the current status instead of statically set it to true on initialization.

* Minor corrections and cleanup

* Refactoring

---------

Co-authored-by: Boris Brock <gabrielborisbrock@gmail.com>